use ms instead of seconds

+++ SoogbadMusic/src/main/java/com/soogbad/soogbadmusic/PlaybackManager.java
@@ -164,7 +164,7 @@ public class PlaybackManager {
         return queue.contains(song);
     }
 
-    public static void setCurrentTime(double time) {
+    public static void setCurrentTime(long time) {
         getPlayer().setCurrentTime(time);
     }
